+++ R/map_greatlakes_fishtissue.R
@@ -22,7 +22,7 @@
 #'
 #' @examples
 #'
-#' data <- data_gl2010
+#' data <- data_gl2010_map
 #' sv <- 40
 #' xlong <- "Longitude"
 #' ylat <- "Latitude"